AD-288 rename event
