+++ analytics-firebase/src/commonMain/kotlin/dev/icerock/moko/firebase/FirebaseAdapter.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright 2021 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.icerock.moko.firebase
+
+interface FirebaseAdapter {
+    fun sendFirebaseEvent(amplitudeEvent: AmplitudeEvent)
+}
