+++ analytics-firebase/src/commonMain/kotlin/dev/icerock/moko/firebase/FirebaseEvent.kt
@@ -1,11 +1,11 @@
 /*
  * Copyright 2021 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
-package dev.icerock.moko.appMetrica
+package dev.icerock.moko.firebase
 
-interface AppMetricaEvent
+interface FirebaseEvent
 
-class SimpleAppMetricaEvent(
+class FirebaseSimpleEvent(
     val id: String? = null,
     val value: String? = null,
-) : AppMetricaEvent
+) : FirebaseEvent
